Cryptic_IP_Binding_Sites/requirements.txt
biopython>=1.81
freesasa>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.28.0
py3Dmol>=2.0.0
ipywidgets>=8.0.0
pytest>=7.3.0
